Improve auditing configuration
