+++ spring-boot-actuator/src/test/java/org/springframework/boot/actuate/autoconfigure/AuditAutoConfigurationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2013 the original author or authors.
+ * Copyright 2012-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,9 +22,12 @@ import org.springframework.boot.actuate.audit.AuditEventRepository;
 import org.springframework.boot.actuate.audit.InMemoryAuditEventRepository;
 import org.springframework.boot.actuate.security.AuthenticationAuditListener;
 import org.springframework.boot.actuate.security.AuthorizationAuditListener;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.event.AbstractAuthorizationEvent;
+import org.springframework.security.authentication.event.AbstractAuthenticationEvent;
 
 import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.Assert.assertNotNull;
@@ -34,16 +37,15 @@ import static org.junit.Assert.assertThat;
  * Tests for {@link AuditAutoConfiguration}.
  *
  * @author Dave Syer
+ * @author Vedran Pavic
  */
 public class AuditAutoConfigurationTests {
 
-	private AnnotationConfigApplicationContext context;
+	private AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
 
 	@Test
 	public void testTraceConfiguration() throws Exception {
-		this.context = new AnnotationConfigApplicationContext();
-		this.context.register(AuditAutoConfiguration.class);
-		this.context.refresh();
+		registerAndRefresh(AuditAutoConfiguration.class);
 		assertNotNull(this.context.getBean(AuditEventRepository.class));
 		assertNotNull(this.context.getBean(AuthenticationAuditListener.class));
 		assertNotNull(this.context.getBean(AuthorizationAuditListener.class));
@@ -51,15 +53,35 @@ public class AuditAutoConfigurationTests {
 
 	@Test
 	public void ownAutoRepository() throws Exception {
-		this.context = new AnnotationConfigApplicationContext();
-		this.context.register(Config.class, AuditAutoConfiguration.class);
-		this.context.refresh();
+		registerAndRefresh(CustomAuditEventRepositoryConfiguration.class,
+				AuditAutoConfiguration.class);
 		assertThat(this.context.getBean(AuditEventRepository.class),
 				instanceOf(TestAuditEventRepository.class));
 	}
 
+	@Test
+	public void ownAuthenticationAuditListener() throws Exception {
+		registerAndRefresh(CustomAuthenticationAuditListenerConfiguration.class,
+				AuditAutoConfiguration.class);
+		assertThat(this.context.getBean(AuthenticationAuditListener.class),
+				instanceOf(TestAuthenticationAuditListener.class));
+	}
+
+	@Test
+	public void ownAuthorizationAuditListener() throws Exception {
+		registerAndRefresh(CustomAuthorizationAuditListenerConfiguration.class,
+				AuditAutoConfiguration.class);
+		assertThat(this.context.getBean(AuthorizationAuditListener.class),
+				instanceOf(TestAuthorizationAuditListener.class));
+	}
+
+	private void registerAndRefresh(Class<?>... annotatedClasses) {
+		this.context.register(annotatedClasses);
+		this.context.refresh();
+	}
+
 	@Configuration
-	public static class Config {
+	public static class CustomAuditEventRepositoryConfiguration {
 
 		@Bean
 		public TestAuditEventRepository testAuditEventRepository() {
@@ -69,6 +91,50 @@ public class AuditAutoConfigurationTests {
 	}
 
 	public static class TestAuditEventRepository extends InMemoryAuditEventRepository {
+	}
+
+	@Configuration
+	protected static class CustomAuthenticationAuditListenerConfiguration {
+
+		@Bean
+		public AuthenticationAuditListener authenticationAuditListener() {
+			return new TestAuthenticationAuditListener();
+		}
+
+	}
+
+	protected static class TestAuthenticationAuditListener implements AuthenticationAuditListener {
+
+		@Override
+		public void setApplicationEventPublisher(ApplicationEventPublisher publisher) {
+		}
+
+		@Override
+		public void onApplicationEvent(AbstractAuthenticationEvent event) {
+		}
+
+	}
+
+	@Configuration
+	protected static class CustomAuthorizationAuditListenerConfiguration {
+
+		@Bean
+		public AuthorizationAuditListener authorizationAuditListener() {
+			return new TestAuthorizationAuditListener();
+		}
+
+	}
+
+	protected static class TestAuthorizationAuditListener implements AuthorizationAuditListener {
+
+		@Override
+		public void setApplicationEventPublisher(ApplicationEventPublisher publisher) {
+		}
+
+		@Override
+		public void onApplicationEvent(AbstractAuthorizationEvent event) {
+		}
+
 	}
 
 }
