+++ spring-boot-docs/src/main/asciidoc/production-ready-features.adoc
@@ -1324,7 +1324,9 @@ channel and any subscribers you need).
 Spring Boot Actuator has a flexible audit framework that will publish events once Spring
 Security is in play ('`authentication success`', '`failure`' and '`access denied`'
 exceptions by default). This can be very useful for reporting, and also to implement a
-lock-out policy based on authentication failures.
+lock-out policy based on authentication failures. To customize published security events
+you can provide your own implementations of `AuthenticationAuditListener` and
+`AuthorizationAuditListener`.
 
 You can also choose to use the audit services for your own business events. To do that
 you can either inject the existing `AuditEventRepository` into your own components and
